resync dev branch with trunk changes
git-svn-id: https://svn.apache.org/repos/asf/incubator/trafficserver/traffic/branches/dev@905849 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cluster/ClusterCache.cc
@@ -2216,7 +2216,7 @@ CacheContinuation::remoteOpEvent(int event_code, Event * e)
     start_time = 0;
   }
   // for CACHE_EVENT_RESPONSE/XXX the lock was acquired at the higher level
-  int return_error = 0;
+  intptr_t return_error = 0;
   ClusterVCToken *pToken = NULL;
 
 retry:
@@ -2945,7 +2945,7 @@ int
 CacheContinuation::callbackEvent(int event, Event * e)
 {
   if (!action.cancelled)
-    action.continuation->handleEvent(result, (void *) result_error);
+    action.continuation->handleEvent(result, (void *)(intptr_t)result_error);
   cacheContAllocator_free(this);
   return EVENT_DONE;
 }
